requirements.txt
aiogram[fast]==3.4.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
